requirements.txt
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-frontmatter>=1.0.0
watchdog>=3.0.0
pydantic>=2.0.0
numpy>=1.24.0
